requirements.txt
PySide6==6.6.1
requests>=2.31.0
python-dotenv==1.0.0
pymongo==4.6.1
bcrypt==4.1.2
python-dateutil>=2.8.2
pillow>=10.0.0
emoji==2.8.0
